pyproject.toml
[build-system]
requires      = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kp-gscapi"
version = "1.0.0"
description = "Custom wrapper for Google Search Console API"
readme = "README.md"
authors = [{ name = "Kenny", email = "hi@anhtrinh.com" }]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords = ["feed", "reader", "tutorial"]
dependencies = [
    "google-api-python-client",
    "pandas",
    
]
requires-python = ">=3.9"

[project.optional-dependencies]

[project.urls]
Homepage = "https://github.com/kp9z"

[project.scripts]
realpython = "reader.__main__:main"

[tool.bumpver]
current_version = "1.0.2"
version_pattern = "MAJOR.MINOR.PATCH"
commit_message = "bump version {old_version} -> {new_version}"
commit = true
tag = true
push = true

[tool.bumpver.file_patterns]
"pyproject.toml" = [
    'current_version = "{version}"',
]


